Fold RTCPeerConnectionBase into RTCPeerConnection and getUserMediaImpl into getUserMedia
