+++ getUserMedia.js
@@ -1,5 +1,18 @@
 'use strict';
 
-var getUserMediaImpl = require('./getUserMediaImpl');
+const WebRTCModule = require('react-native').NativeModules.WebRTCModule;
+const MediaStream = require('./MediaStream');
+const MediaStreamTrack = require('./MediaStreamTrack');
 
-module.exports = getUserMediaImpl;
+function getUserMedia(constraints, success, failure) {
+  WebRTCModule.getUserMedia(constraints, (id, tracks) => {
+    const stream = new MediaStream(id);
+    for (var i = 0; i < tracks.length; i++) {
+      stream.addTrack(new MediaStreamTrack(tracks[i]));
+    }
+
+    success(stream);
+  });
+}
+
+module.exports = getUserMedia;
